+++ src/attached_probe.cpp
@@ -30,6 +30,7 @@
 #include "log.h"
 #include "probe_matcher.h"
 #include "usdt.h"
+#include "utils.h"
 
 namespace bpftrace {
 
@@ -758,9 +759,13 @@ void AttachedProbe::load_prog(BPFfeature &feature)
         opts.expected_attach_type = static_cast<::bpf_attach_type>(
             libbpf::BPF_TRACE_ITER);
 
+      // We want to avoid kprobe_multi when a module is specified
+      // because the BPF_TRACE_KPROBE_MULTI link type does not
+      // currently support the `module:function` syntax.
       if ((probe_.type == ProbeType::kprobe ||
            probe_.type == ProbeType::kretprobe) &&
-          feature.has_kprobe_multi() && !probe_.funcs.empty())
+          feature.has_kprobe_multi() && !probe_.funcs.empty() &&
+          probe_.path.empty())
         opts.expected_attach_type = static_cast<::bpf_attach_type>(
             libbpf::BPF_TRACE_KPROBE_MULTI);
 
@@ -960,11 +965,39 @@ void AttachedProbe::attach_kprobe(bool safe_mode)
     return;
   }
 
+  // Construct a string containing "module:function."
+  // Also log a warning or throw an error if the module doesn't exist,
+  // before attempting to attach.
+  // Note that we do not pass vmlinux, if it is specified.
+  std::string funcname = probe_.attach_point;
+  const std::string &modname = probe_.path;
+  if ((modname.length() > 0) && modname != "vmlinux")
+  {
+    if (!is_module_loaded(modname))
+    {
+      std::string message = "specified module " + modname + " in probe " +
+                            probe_.name + " is not loaded.";
+      if (probe_.orig_name != probe_.name)
+      {
+        // Wildcard usage just gets a warning
+        LOG(WARNING) << message;
+      }
+      else
+      {
+        // Explicitly specified modules should fail
+        LOG(ERROR) << message;
+        throw std::runtime_error("Error attaching probe: '" + probe_.name +
+                                 "'");
+      }
+    }
+    funcname = modname + ":" + funcname;
+  }
+
   resolve_offset_kprobe(safe_mode);
   int perf_event_fd = bpf_attach_kprobe(progfd_,
                                         attachtype(probe_.type),
                                         eventname().c_str(),
-                                        probe_.attach_point.c_str(),
+                                        funcname.c_str(),
                                         offset_,
                                         0);