+++ tests/parser.cpp
@@ -1872,7 +1872,7 @@ TEST(Parser, long_param_overflow)
 TEST(Parser, empty_arguments)
 {
   test_parse_failure("::k::vfs_open:: { 1 }");
-  test_parse_failure("k:vfs_open: { 1 }");
+  test_parse_failure("k:vfs_open:: { 1 }");
   test_parse_failure(":w:0x10000000:8:rw { 1 }");
 }